Indev b2.6.2, embedify considered as complete.
